Refactor user management routes and controller
