+++ server/src/modules/usermanagement/usercontroller.js
@@ -178,139 +178,3 @@ exports.update = async (req, res, next) => {
       .json({ message: "An error occurred", error: error.message });
   }
 };
-
-const sendOTP = async (email) => {
-  try {
-    const checkuser = await User.findOne({ email: email });
-    if (!checkuser) {
-      console.log("User not exists");
-      return {
-        success: false,
-        message: "User not exists",
-      };
-    }
-
-    let result = await OTP.findOne({ email });
-    var otp = null;
-    if (result) {
-      otp = result.otp;
-      console.log("OTP already exists:", otp);
-    } else {
-      otp = otpGenerator.generate(6, {
-        lowerCaseAlphabets: false,
-        upperCaseAlphabets: false,
-        specialChars: false,
-      });
-      await OTP.create({ email, otp });
-      console.log("New OTP generated:", otp);
-    }
-
-    console.log(otp);
-    const otpInfo = {
-      title: "Email verification for NITJ",
-      purpose:
-        "Thank you for registering with NITJ. To complete your registration, please use the following OTP (One-Time Password) to verify your account:",
-      OTP: otp,
-    };
-
-    const otpBody = fs.readFileSync(ejsTemplatePath, "utf-8");
-    const renderedHTML = ejs.render(otpBody, otpInfo);
-
-    // Add await here
-    await mailSender(email, "Sign Up verification", renderedHTML);
-
-    return {
-      success: true,
-      message: "OTP sent successfully",
-    };
-  } catch (e) {
-    console.log("Error in sending OTP ", e);
-    return {
-      success: false,
-      message: "Error in sending OTP",
-    };
-  }
-};
-exports.forgotPassword = async (req, res) => {
-  try {
-    const { email } = req.body;
-    const checkuser = await User.findOne({ email: email });
-    if (!checkuser) {
-      console.log("User not exists");
-      return res.status(200).json({
-        success: false,
-        message: "User not exists",
-      });
-    }
-
-    // Generate and send OTP
-    const otp = await sendOTP(email);
-
-    return res.status(200).json({
-      success: true,
-      message: "OTP sent successfully",
-      otp: otp, // Send the OTP to the client for verification
-    });
-  } catch (e) {
-    console.log("Error in sending OTP ", e);
-    return res.status(402).json({
-      success: false,
-      message: "Error in sending OTP",
-    });
-  }
-};
-const verifyOTP = async (email, enteredOTP) => {
-  try {
-    const otpRecord = await OTP.findOne({ email });
-
-    if (!otpRecord) {
-      console.log("No OTP record found for the user");
-      return false;
-    }
-
-    const storedOTP = otpRecord.otp;
-
-    // Compare the entered OTP with the stored OTP
-    return enteredOTP === storedOTP;
-  } catch (error) {
-    console.log("Error verifying OTP: ", error);
-    return false;
-  }
-};
-
-// Reset Password
-exports.resetPassword = async (req, res) => {
-  try {
-    const { email, otp, newPassword } = req.body;
-
-    // Verify the OTP
-    const isOTPValid = await verifyOTP(email, otp);
-
-    if (!isOTPValid) {
-      return res.status(400).json({
-        success: false,
-        message: "Invalid OTP",
-      });
-    }
-
-    // Update the password
-    const hashedPassword = await bcrypt.hash(newPassword, 10);
-    const user = await User.findOneAndUpdate(
-      { email: email },
-      { $set: { password: hashedPassword } },
-      { new: true }
-    );
-
-    return res.status(200).json({
-      success: true,
-      message: "Password updated successfully",
-      user,
-    });
-  } catch (e) {
-    console.log("Error in resetting password ", e);
-    return res.status(500).json({
-      success: false,
-      message: "Error in resetting password",
-    });
-  }
-};
